Change types for currencies to BigDecimal

+++ transaction-service/src/main/java/cz/muni/fi/obs/api/CurrencyExchangeResult.java
@@ -1,12 +1,14 @@
 package cz.muni.fi.obs.api;
 
+import java.math.BigDecimal;
+
 import lombok.Builder;
 
 @Builder
 public record CurrencyExchangeResult(
 		String symbolFrom,
 		String symbolTo,
-		long exchangedRate,
-		long sourceAmount,
-		long exchangedAmount) {
+		BigDecimal exchangedRate,
+		BigDecimal sourceAmount,
+		BigDecimal exchangedAmount) {
 }
